backend/requirements.txt
# Web Framework
Django==5.0.1

# REST Framework
djangorestframework==3.14.0

# API Documentation
drf-spectacular==0.27.1

# Environment Variables
python-dotenv==1.0.1


# CORS Headers
django-cors-headers==4.3.1

# Development Tools
django-debug-toolbar==4.2.0

# Security
django-filter==23.5

# Testing
pytest==7.4.4
pytest-django==4.7.0

# Code Formatting
black==23.12.1
flake8==7.0.0

